# Backend/Services/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "pokemon API",
    "version": "1.0.0",
    "description": "Een eenvoudige API voor pokemone"
  },
  "servers": [
    {
      "url": "http://localhost:3333"
    }
  ],
  "paths": {
    "/Getpokemon": {
      "get": {
        "summary": "Toon alle Pokemon",
        "responses": {
          "200": {
            "description": "Lijst van Pokemon"
          }
        }
      }
    },
    "/pokemon": {
      "get": {
        "summary": "Toon één Pokemon",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Pokemon gevonden" },
          "404": { "description": "Niet gevonden" }
        }
      }
    },
    "/NewPokemon": {
      "post": {
        "summary": "Voeg een pokemon toe",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "name": { "type": "string" },
                  "type": { "type": "string"},
                  "gigantamax": { "type": "string" }
                },
                "required": [ "name", "type", "gigantamax"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Pokemon toegevoegd" }
        }
      }
    },
    "/updatePokemon/{id}": {
      "put": {
        "summary": "Wijzig een pokemon",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "name": { "type": "string" },
                  "type": { "type": "string"},
                  "gigantamax": { "type": "string" }
                },
                "required": ["name", "type", "gigantamax"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Pokemone bijgewerkt" }
        }
      }
    },
    "/deletePokemon/{id}": {
      "delete": {
        "summary": "Verwijder een pokemon",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Verwijderd" },
          "404": { "description": "Niet gevonden" }
        }
      }
    }
  }
}